fix: #SKFP-139 no files without particpants or biospecimen
